examples/authorization-for-tools/llama-index-examples/pyproject.toml
[tool.poetry]
name = "llama-index-examples"
version = "0.1.0"
description = ""
authors = ["Auth0 <support@auth0.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.scripts]
fga_init = "scripts.fga_init:main"
start = "src.main:main"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.1"
questionary = "^2.1.0"
llama-index = "^0.12.49"
llama-index-agent-openai = "^0.4.12"
auth0-ai-llamaindex = { path = "../../../packages/auth0-ai-llamaindex", develop = true }

[tool.poetry.group.dev.dependencies]
openfga-sdk = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
